Added parallel build feature
Added parallel build feature to facilitate faster build times.

By default this feature is disabled. To enable, simply change the
"parallel_builds" variable in your Build Gear configuration file to eg. a
value of 2. This means that maximum two builds can be building at the same
time.

Known issue: The build log becomes a mess when building in parallel
because it is not protected against simultaneous log writes from the
various builds running in parallel (TO BE FIXED).

This solution cleans out the old thread implementation in favour of the
new c++11 threading library. Consequently, a c++11 enabled C++ compiler
is now required. Generally, this results in a much cleaner implementation.