Fixed build log handling for parallel builds
Previously, when building in parallel, the build log got messed up
because each build process wrote to the build log file at the same time.
The end result was a very incoherent build log file.

This is now fixed so that each build process only dumps its log output
to the build log file when it is finished. Also, the build log file is
only written to by one build process at a time.
The end result is a bursting but coherent log.