Gracefully exit if build fails
buildgear.sh is made leader of a process group when it is called.
The script will communicate its PID to buildgear using a named fifo.

This allows buildgear to kill all build scripts, and their children
if a build fails.

This is done in order to ensure that the log of the failing build
will be written to the build log.