Added layer support
A concept of layers is introduced to buildgear. This will help in situations
where a specific board needs a different configuration in the buildfile of
a build, that is already in the mini distribution.

A prioritized array of layer names are given in the buildfiles/config file i.e.
LAYERS=(myLayer1 default)
In this example, myLayer1 overrides the default layer.

To make a particular Buildfile override any other build of same name, simply include a
.layer file in the Buildfile directory.

The .layer file should contain the name of the layer that the build should belong to. i.e.
myLayer1

If no .layer file is present, or a layer name, that is not given in the LAYER_PRIORITY array,
a layer called 'default' is used.

The terminal output during build, and the build log, will show the layer name of a build, if
it is not on the default layer.