Added config command
A new config command has been added. The config command can be used
to change the buildgear configuration file through the commandline.

A --global flag can be used to change the configuration in the global
buildgear configuration file.

A --unset flag can be used to revert the configuration to default

Ex.
config parallel_builds 4
will set the parallel_builds in the local buildgear configuration file
to 4.