Improved configfile handling
The config part has been rewritten a bit. Options that are defined in
the buildgear config files are stored un an unordered_map i.e.
bg_config["parallel_builds"]. This allows for a more compact code base.

The config command will now update the config files with the installed
templates when a option is changed. This will make sure that the config
files are up-to-date if more options are introduced in the future.