Introduce lock file for the build root
A lock file is used to ensure that multiple instances of buildgear
are not allowed to run, if they have the possibility of changing data.

I.e. you cannot run two build commands in the same root dir, but you are
allowed to run a build command and a show --log commmand.