Added a load chart feature
If a --load-chart option is given to the build command, the CPU load
and memory usage will be recorded throughout the build. When the build
is finished the loads are rendered in a svg file that shows the results.

The svg with the data is located in build/output/load-chart.<build name>.svg