Add options to set custom ssh key locations
The location of the public/private SSH keys used for SFTP/SCP download
can be configured with buildgear config.

The location of the public key is set with the keyword ssh_public_key.
The location of the private key is set with the keyword ssh_private_key.

Defaults are $HOME/.ssh/id_rsa(.pub)