Added support for "build-lock" option
This feature adds support for locking a build so that it will be the
only one building at any given time (even when parallel_builds > 1).
This is done by introducing a new "options" list and a special toggle
option named "build-lock".

To force a buildfile to be built in non-parallel, simply add to the
buildfile:

options=build-lock

This is reported to be useful for eg. kernel module builds which must
avoid to be built in parallel at the same time because they need to
operate in the same kernel build workspace in order to produce coherent
kernel and modules.

In the future, more options can be added to the options list.