Automatic assigment of layer by buildfile location
Buildfiles are now automatically assigned to <layer> when placed in:

buildfiles/cross/cross-<layer>
buildfiles/native/native-<layer>

Provided that <layer> is defined in the master config LAYERS list.

If <layer> is not defined in the LAYERS list buildfiles in these
locations will continue to default to the "default" layer.

However, the .layer file will still take priority if used.

This feature makes it much more easy to assign multiple buildfiles to a
single layer by not having to create a .layer file for each buildfile
but instead by simply putting them in the right layer directory
location.