Removed '.layer' file support
Defining which layer a Buildfile belongs to using the .layer file is now
deprecated.

Instead the layer can now be defined in the Buildfile itself by adding
the "layer" variable. For example, to assign a Buildfile to "mylayer"
simply add:

layer=mylayer

If the layer is not defined in the master config LAYERS list, then the
Buildfile will default to the "default" layer.

Please note that the preferred method for assigning layers is by
Buildfile path location (eg. cross-<layer>) and not by using the "layer"
variable.

This is part of a plan to consolidate all settings inside the Buildfile
to make the Buildfile more precise to avoid that the user misses any
settings.