Removed '.nostrip' file support
Defining whether or not a Buildfile is stripped or not using the
.nostrip file is now deprecated.

Instead the nostrip feature can now be defined in the Buildfile itself
in either of two ways:

1. Disable strip entirely

   Adding a 'nostrip' option to the 'options' list like so:

   options=(nostrip)

   Will disable any strip of the build entirely.

2. Selective nostrip

   It is possible to select which files will not be stripped by adding a
   nostrip array. For example:

   nostrip=(/usr/lib/libtest.so /usr/lib/libtest.a)

   Entries in the nostrip array are treated as regular expressions (as
   used by grep).

Note: (2) will have no effect if (1) is used.

Note: By default all executables and .so and .a files are stripped.

This change is part of a plan to consolidate all settings inside the
Buildfile to make the Buildfile more precise to avoid that the user
overlooks any settings.