Add rebuild detection based on buildfile checksum
This means that a build will automatically rebuild if the sha256sum
checksum of its bash expanded Buildfile changes.

This is particular useful if you use configurations variables to
reconfigure the way your firmware is built.

For example, if you have a configuration variable in the global
buildfiles configuration file (config/config), eg.:

CONFIG_AUDIO="alsa"

And in one or more Buildfiles you use this variable to reconfigure the
build or change dependencies. For example:

if [ "$CONFIG_AUDIO" == "alsa" ]
then
   CONF_OPTIONS+="--audio=alsa"
   depends+=(alsa-lib)
else
   CONF_OPTIONS+="--audio=oss"
fi

Now, if you change the value of CONFIG_AUDIO between builds, then build
gear will automatically rebuild any buildfiles that uses CONFIG_AUDIO
because it will detect that the the checksum of the bash expanded
buildfile(s) changes by this modification.